More dynamic text, star generation!
Randomly generate stars. Currently is a mix of manually created stars and random ones, but still figuring out which to keep or if I would like to keep them both to maintain more even distribution. Also made them smaller.
In addition, changed text size to be in terms of vw rather than px in the "description" section. Still deciding how to proceed with that part, which I plan to maybe have a typewriter effect.